refactor: test 코드 내 fixture 위치 이동
