+++ src/test/java/com/example/hot_deal/member/service/MemberServiceTest.java
@@ -5,7 +5,7 @@ import com.example.hot_deal.member.domain.entity.Member;
 import com.example.hot_deal.member.domain.repository.MemberRepository;
 import com.example.hot_deal.member.dto.RegisterRequest;
 import com.example.hot_deal.member.dto.RegisterResponse;
-import com.example.hot_deal.member.fixture.MemberFixture;
+import com.example.hot_deal.fixture.MemberFixture;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
